[#75] Adding the new logic to a new helper
Created the relationalFields.js helper
