+++ src/helpers/database/index.js
@@ -1,9 +1,11 @@
 const dbHelpers = require('./dbHelpers');
 const queries = require('./queries');
 const events = require('./events');
+const relationalFields = require('./relationalFields');
 
 module.exports = {
     dbHelpers,
     queries,
-    events
+    events,
+    relationalFields
 };
